Dodanie pól do Advert (tytuł, data utworzenia, data wygaśnęcia, czy zaakceptowane), zmiana nazwy MainPageController na AdvertController, zmiana ścieżek w AdertController, dodanie usuwania ogłoszeń

+++ src/main/java/com/example/projektJava/dao/AdvertDAO.java
@@ -8,7 +8,7 @@ public interface AdvertDAO {
 
     void save(Advert theAdvert);
 
-    Advert findById(Integer id);
+    Advert findById(Long id);
 
     List<Advert> findAll();
